Move prettyPrintKast from kast to ktool.kprint
