+++ pyk/src/pyk/ktool/__init__.py
@@ -1,3 +1,14 @@
 from .kompile import KompileBackend, kompile
-from .kprint import KPrint
+from .kprint import (
+    KPrint,
+    appliedLabelStr,
+    binOpStr,
+    buildSymbolTable,
+    indent,
+    newLines,
+    paren,
+    prettyPrintKast,
+    prettyPrintKastBool,
+    underbarUnparsing,
+)
 from .kprove import KProve, kprovex
